+++ CHANGELOG.md
@@ -7,6 +7,11 @@ CHANGELOG
 * **Breaking** Updated `TransactionReport` to make the `ipAddress` parameter
   optional. Now the `tag` and at least one of the following paramters must be
   supplied: `ipAddress`, `maxmindId`, `minfraudId`, `transactionId`.
+* Added `billingPhone` and `shippingPhone` properties to the minFraud Insights
+  and Factors response models. These contain objects with information about
+  the respective phone numbers. Please see [our developer
+  site](https://dev.maxmind.com/minfraud/api-documentation/responses/) for
+  more information.
 
 6.1.0 (2024-04-16)
 ------------------